+++ cryptobackend/internal/fips140state/systemfips_openssl.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//go:build goexperiment.systemcrypto
+//go:build goexperiment.systemcrypto && (linux || freebsd)
 
 package fips140state
